# Conflicts:
#	book-your-lesson/pom.xml
#	book-your-lesson/src/main/java/ro/bydl/cfg/AplicationConfiguration.java
#	book-your-lesson/src/main/java/ro/bydl/controler/ScheduleControler.java
#	book-your-lesson/src/main/java/ro/bydl/dao/database/JdbcTemplateScheduleDAO.java
#	book-your-lesson/src/main/java/ro/bydl/dao/database/JdbcTemplateScheduleDao.java
#	book-your-lesson/src/main/java/ro/bydl/dao/database/JdbcTemplateScheduleeDao.java
#	book-your-lesson/src/main/java/ro/bydl/service/CalendarHelper.java
#	book-your-lesson/src/main/java/ro/bydl/service/PersonHelper.java
#	book-your-lesson/src/main/java/ro/bydl/service/PersonService.java
#	book-your-lesson/src/main/java/ro/bydl/service/ScheduleService.java
#	book-your-lesson/src/main/java/ro/bydl/service/StudentHelper.java
#	book-your-lesson/src/main/java/ro/bydl/service/StudentService.java
#	book-your-lesson/src/main/java/ro/bydl/service/TeacherService.java
#	book-your-lesson/src/test/java/ro/bydl/IMStudentDAO.java
#	book-your-lesson/target/classes/ro/bydl/controler/ScheduleControler.class
#	book-your-lesson/target/classes/ro/bydl/service/ScheduleService.class

+++ book-your-lesson/src/main/java/ro/bydl/domain/Schedule.java
@@ -10,12 +10,12 @@ public class Schedule extends AbstractModel {
 	private String date;
 	private int week;
 	private boolean isFree;
-	private int studentId;
-	private int teacherId;
+	private long studentId;
+	private long teacherId;
 	@NotBlank
 	private String status;
 	
-	public int getStudentId() {
+	public long getStudentId() {
 		return studentId;
 	}
 
@@ -23,7 +23,7 @@ public class Schedule extends AbstractModel {
 		this.studentId = studentId;
 	}
 
-	public int getTeacherId() {
+	public long getTeacherId() {
 		return teacherId;
 	}
 
@@ -89,7 +89,7 @@ public class Schedule extends AbstractModel {
 	@Override
 	public int hashCode() {
 		final int prime = 31;
-		int result = 1;
+		long result = 1;
 		result = prime * result + ((date == null) ? 0 : date.hashCode());
 		result = prime * result + endHour;
 		result = prime * result + (isFree ? 1231 : 1237);
@@ -98,7 +98,7 @@ public class Schedule extends AbstractModel {
 		result = prime * result + studentId;
 		result = prime * result + teacherId;
 		result = prime * result + week;
-		return result;
+		return (int) result;
 	}
 
 	@Override
